requirements.txt
numpy==1.18.1
urllib3==1.25.9
flask==1.1.2
Flask-PyMongo==2.3.0
opencv-python==4.4.0.44
pandas==1.1.3
scikit-image==0.16.2
scipy==1.4.1
pillow==8.0.0
imagehash==4.1.0
jsonpickle==1.4.1